Refactored AutoComplete Box and fixed styling a bit
